backend/requirements.txt
# Hair Omakase Backend Dependencies
# Python 3.12 / Conda: mini-hair

# Web Framework
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.21
starlette==0.50.0

# Validation & Config
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1

# HTTP Client
requests==2.32.5
httpx==0.28.1

# Google Gemini AI
google-genai==1.56.0

# Image Processing
pillow==12.0.0